Force example group into the rspec-puppet "type" group
Previously the test failed when run against puppet6, because the
mailalias dependency is no longer in puppet, so
`Puppet::Type.type(:mailalias)` returned nil.

To ensure that the module is using the `mailalias` fixture, defined in
.fixtures.yml, we need to rely on RSpec's automagic behavior that
configures puppet.

To do that, our examples need to be in one of rspec-puppet's predefined
example group categories[1]. We could move the spec to a predefined
directory structure, or just make the `type` explicit. As a result of
this change, the spec/fixtures/modules directory will be added to
puppet's modulepath so that the autoloader can load the `mailalias`
type.

[1] https://github.com/rodjek/rspec-puppet/#example-groups